Change FatalsToBrowser to pad error messages.
FatalsToBrowser messages need to padded to over 512 bytes even when not under mod_perl.
MSIE is hard to identify reliably, so pad all messages.  The cost of padding the messages should be slight and not worth the bother of optimizing to a browser, especially as FatalsToBrowser is generally used over local connections.